refactor(userspace): extract ProcessStatus/runtime status to protocol_status.go
Seventh code-motion increment of the protocol.go wire-protocol split
(#4839). Moves the top-level ProcessStatus aggregator (the large helper
status payload), its custom JSON (un)marshalers, and the process-wide
runtime status types into a cohesive sibling protocol_status.go in the
same userspace package.

Moved verbatim (no rename, no logic or json-tag change): ProcessStatus,
its MarshalJSON/UnmarshalJSON methods, WorkerRuntimeStatus,
SlowPathStatus, PacketResolution, FlowTupleStatus, FlowWorkerStatus,
ExceptionStatus.

The encoding/json and time imports migrate with the ProcessStatus
methods and time.Time fields; protocol.go no longer needs
encoding/json. Behavior byte-identical, including the ProcessStatus
value-receiver MarshalJSON legacy-alias emission. Build passes; gofmt
clean.
